requirements.txt
asgiref==3.8.1
Django==5.1.6
django-debug-toolbar==5.0.1
psycopg-binary==3.2.4
psycopg2==2.9.10
sqlparse==0.5.3
tzdata==2025.1
